XCODE_RUNNING_FOR_PREVIEWSフラグを使用して自動的にプレビュー用DIに切り替える仕組みを実装
