+++ MediaLibrary/Sources/DependencyInjection/AppContainer.swift
@@ -9,11 +9,24 @@ package enum AppDependencies {
     // MARK: - Infrastructure Layer
 
     /// MediaRepository の実装
-    package static let mediaRepository: any MediaRepository = MediaRepositoryImpl()
+    package static let mediaRepository: any MediaRepository = {
+        #if DEBUG
+        if ProcessInfo.processInfo.environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1" {
+            return PreviewDependencies.mediaRepository
+        }
+        #endif
+        return MediaRepositoryImpl()
+    }()
 
     /// PhotoLibraryPermissionService の実装
-    package static let photoLibraryPermissionService: any PhotoLibraryPermissionService =
-        PhotoLibraryPermissionServiceImpl()
+    package static let photoLibraryPermissionService: any PhotoLibraryPermissionService = {
+        #if DEBUG
+        if ProcessInfo.processInfo.environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1" {
+            return PreviewDependencies.photoLibraryPermissionService
+        }
+        #endif
+        return PhotoLibraryPermissionServiceImpl()
+    }()
 
     // MARK: - Application Layer
 
@@ -23,6 +36,84 @@ package enum AppDependencies {
         permissionService: photoLibraryPermissionService
     )
 }
+
+// MARK: - Preview Support
+
+#if DEBUG
+/// プレビュー用のモックリポジトリ
+fileprivate struct MockMediaRepository: MediaRepository {
+    func fetchMedia() async throws -> [Media] {
+        return try [
+            Media(
+                id: Media.ID("1"),
+                type: .photo,
+                metadata: Media.Metadata(format: .jpeg, capturedAt: Date()),
+                filePath: "dummy1.jpg"
+            ),
+            Media(
+                id: Media.ID("2"),
+                type: .photo,
+                metadata: Media.Metadata(format: .jpeg, capturedAt: Date()),
+                filePath: "dummy2.jpg"
+            ),
+            Media(
+                id: Media.ID("3"),
+                type: .photo,
+                metadata: Media.Metadata(format: .png, capturedAt: Date()),
+                filePath: "dummy3.png"
+            ),
+        ]
+    }
+    
+    func fetchThumbnail(for mediaID: Media.ID, size: CGSize) async throws -> Media.Thumbnail {
+        #if canImport(UIKit)
+        import UIKit
+        
+        let systemImages = ["photo.fill", "camera.fill", "video.fill", "heart.fill", "star.fill"]
+        let colors: [UIColor] = [.systemBlue, .systemGreen, .systemOrange, .systemPink, .systemPurple]
+        
+        let index = abs(mediaID.value.hashValue) % systemImages.count
+        let imageName = systemImages[index]
+        let color = colors[index]
+        
+        let config = UIImage.SymbolConfiguration(pointSize: min(size.width, size.height) * 0.6, weight: .medium)
+        let image = UIImage(systemName: imageName, withConfiguration: config)?.withTintColor(color, renderingMode: .alwaysOriginal)
+        
+        let data = image?.jpegData(compressionQuality: 0.8) ?? UIImage().jpegData(compressionQuality: 0.8)!
+        return try Media.Thumbnail(
+            mediaID: mediaID,
+            imageData: data,
+            size: size
+        )
+        #else
+        // macOS等でUIKitが使えない場合のフォールバック
+        let data = Data()
+        return try Media.Thumbnail(
+            mediaID: mediaID,
+            imageData: data,
+            size: size
+        )
+        #endif
+    }
+}
+
+/// プレビュー用のモック権限サービス
+fileprivate struct MockPhotoLibraryPermissionService: PhotoLibraryPermissionService {
+    func checkPermissionStatus() -> PhotoLibraryPermissionStatus {
+        return .authorized
+    }
+    
+    func requestPermission() async -> PhotoLibraryPermissionStatus {
+        return .authorized
+    }
+}
+
+/// プレビュー用の依存関係を管理する構造体
+package enum PreviewDependencies {
+    package static let mediaRepository: any MediaRepository = MockMediaRepository()
+    package static let photoLibraryPermissionService: any PhotoLibraryPermissionService = MockPhotoLibraryPermissionService()
+}
+#endif
 
 // MARK: - Test Support
